Factoring out checks
