+++ frontend/src/components/dashboard/DashboardTransactionsHistoryChart.tsx
@@ -6,7 +6,7 @@ import moment from "moment";
 import PaginationSpinner from "../utils/PaginationSpinner";
 
 import { useTranslation } from "react-i18next";
-import { useChainTransactionStats } from "../../hooks/subscriptions";
+import { TransactionCountHistory } from "../../libraries/wamp/types";
 
 const chartsStyle = {
   height: 232,
@@ -14,119 +14,121 @@ const chartsStyle = {
   marginBottom: 26,
 };
 
-const DashboardTransactionHistoryChart: React.FC = React.memo(() => {
-  const { t } = useTranslation();
-  const transactionsCountHistoryForTwoWeeks =
-    useChainTransactionStats()?.transactionsCountHistoryForTwoWeeks || [];
+type Props = {
+  transactionHistory?: TransactionCountHistory[];
+};
 
-  const getDate = () => {
-    const format = t(
-      "component.dashboard.DashboardTransactionHistoryChart.date_format"
-    );
-    const date = transactionsCountHistoryForTwoWeeks.map((t) =>
-      moment(t.date).format(format)
-    );
-    return date;
-  };
+const DashboardTransactionHistoryChart: React.FC<Props> = React.memo(
+  ({ transactionHistory = [] }) => {
+    const { t } = useTranslation();
 
-  const count = transactionsCountHistoryForTwoWeeks.map((t) => t.total);
-  const getOption = () => {
-    return {
-      title: {
-        text: t(
-          "component.dashboard.DashboardTransactionHistoryChart.14_day_history.title"
-        ),
-      },
-      tooltip: {
-        trigger: "axis",
-        position: "top",
-        backgroundColor: "#25272A",
-        formatter: `{b0}<br />${t(
-          "component.dashboard.DashboardTransactionHistoryChart.14_day_history.transactions"
-        )}: {c0}`,
-      },
-      grid: {
-        left: "5%",
-        bottom: "3%",
-        containLabel: true,
-        backgroundColor: "#F9F9F9",
-        show: true,
-        color: "white",
-        borderWidth: 0,
-      },
-      xAxis: [
-        {
-          type: "category",
-          boundaryGap: false,
-          data: getDate(),
-          axisLine: {
-            show: false,
-          },
-          axisLabel: {
-            color: "#9B9B9B",
-          },
-          offset: 3,
-          axisTick: {
-            show: false,
-          },
+    const getDate = () => {
+      const format = t(
+        "component.dashboard.DashboardTransactionHistoryChart.date_format"
+      );
+      const date = transactionHistory.map((t) => moment(t.date).format(format));
+      return date;
+    };
+
+    const count = transactionHistory.map((t) => t.total);
+    const getOption = () => {
+      return {
+        title: {
+          text: t(
+            "component.dashboard.DashboardTransactionHistoryChart.14_day_history.title"
+          ),
         },
-      ],
-      yAxis: [
-        {
-          type: "value",
-          splitLine: {
-            lineStyle: {
-              color: "white",
+        tooltip: {
+          trigger: "axis",
+          position: "top",
+          backgroundColor: "#25272A",
+          formatter: `{b0}<br />${t(
+            "component.dashboard.DashboardTransactionHistoryChart.14_day_history.transactions"
+          )}: {c0}`,
+        },
+        grid: {
+          left: "5%",
+          bottom: "3%",
+          containLabel: true,
+          backgroundColor: "#F9F9F9",
+          show: true,
+          color: "white",
+          borderWidth: 0,
+        },
+        xAxis: [
+          {
+            type: "category",
+            boundaryGap: false,
+            data: getDate(),
+            axisLine: {
+              show: false,
+            },
+            axisLabel: {
+              color: "#9B9B9B",
+            },
+            offset: 3,
+            axisTick: {
+              show: false,
             },
           },
-          splitNumber: 3,
-          axisLine: {
-            show: false,
-          },
-          axisLabel: {
-            color: "#9B9B9B",
-          },
-          offset: 3,
-          axisTick: {
-            show: false,
-          },
-        },
-      ],
-      series: [
-        {
-          name: "Txns",
-          type: "line",
-          smooth: true,
-          lineStyle: {
-            color: "#00C1DE",
-            width: 2,
-          },
-          symbol: "circle",
-          itemStyle: {
-            color: "#25272A",
-          },
-          areaStyle: {
-            color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
-              {
-                offset: 0,
-                color: "rgba(0, 193, 222, 0.19)",
+        ],
+        yAxis: [
+          {
+            type: "value",
+            splitLine: {
+              lineStyle: {
+                color: "white",
               },
-              {
-                offset: 1,
-                color: "rgba(197, 247, 255, 0)",
-              },
-            ]),
+            },
+            splitNumber: 3,
+            axisLine: {
+              show: false,
+            },
+            axisLabel: {
+              color: "#9B9B9B",
+            },
+            offset: 3,
+            axisTick: {
+              show: false,
+            },
           },
-          data: count,
-        },
-      ],
+        ],
+        series: [
+          {
+            name: "Txns",
+            type: "line",
+            smooth: true,
+            lineStyle: {
+              color: "#00C1DE",
+              width: 2,
+            },
+            symbol: "circle",
+            itemStyle: {
+              color: "#25272A",
+            },
+            areaStyle: {
+              color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
+                {
+                  offset: 0,
+                  color: "rgba(0, 193, 222, 0.19)",
+                },
+                {
+                  offset: 1,
+                  color: "rgba(197, 247, 255, 0)",
+                },
+              ]),
+            },
+            data: count,
+          },
+        ],
+      };
     };
-  };
 
-  if (transactionsCountHistoryForTwoWeeks.length === 0) {
-    return <PaginationSpinner />;
+    if (transactionHistory.length === 0) {
+      return <PaginationSpinner />;
+    }
+    return <ReactEcharts option={getOption()} style={chartsStyle} />;
   }
-  return <ReactEcharts option={getOption()} style={chartsStyle} />;
-});
+);
 
 export default DashboardTransactionHistoryChart;
