+++ frontend/src/hooks/subscriptions.ts
@@ -3,8 +3,11 @@ import { useWampSubscription } from "./wamp";
 export const useChainBlockStats = () =>
   useWampSubscription("chain-blocks-stats");
 
-export const useChainTransactionStats = () =>
-  useWampSubscription("chain-transactions-stats");
+export const useRecentTransactions = () =>
+  useWampSubscription("recent-transactions");
+
+export const useTransactionHistory = () =>
+  useWampSubscription("transaction-history");
 
 export const useFinalityStatus = () => useWampSubscription("finality-status");
